Feat/category list sort
