+++ src/components/category/Game.jsx
@@ -14,22 +14,24 @@ const Game = ({ code, posts }) => {
       </St.CategoryHeader>
       <St.WriteBtn to={`/write`}>글 작성하기</St.WriteBtn>
       <St.PostListWrap>
-        {categoryPosts.map((post) => (
-          <St.PostList key={post.id} to={`/detail/${post.id}`}>
-            <St.PostTime>{dayjs(post.created_at).format('YYYY-MM-DD')}</St.PostTime>
-            <St.PostTitle key={post.id}>{post.title}</St.PostTitle>
-            <St.PostRight>
-              <St.ListLikeBox>
-                <FaGlassCheers size="25" color="#eea100" />
-                <p>{post.likes.length}</p>
-              </St.ListLikeBox>
-              <St.ListProfileImgBox>
-                <St.ListProfileImg alt="이미지 준비중" src={post.profileImg}></St.ListProfileImg>
-              </St.ListProfileImgBox>
-              <St.PostUser>{post.nickname}</St.PostUser>
-            </St.PostRight>
-          </St.PostList>
-        ))}
+        {categoryPosts
+          .sort((a, b) => dayjs(b.created_at) - dayjs(a.created_at))
+          .map((post) => (
+            <St.PostList key={post.id} to={`/detail/${post.id}`}>
+              <St.PostTime>{dayjs(post.created_at).format('YYYY-MM-DD')}</St.PostTime>
+              <St.PostTitle key={post.id}>{post.title}</St.PostTitle>
+              <St.PostRight>
+                <St.ListLikeBox>
+                  <FaGlassCheers size="25" color="#eea100" />
+                  <p>{post.likes.length}</p>
+                </St.ListLikeBox>
+                <St.ListProfileImgBox>
+                  <St.ListProfileImg alt="이미지 준비중" src={post.profileImg}></St.ListProfileImg>
+                </St.ListProfileImgBox>
+                <St.PostUser>{post.nickname}</St.PostUser>
+              </St.PostRight>
+            </St.PostList>
+          ))}
       </St.PostListWrap>
     </div>
   );
